requirements.txt
# Core AI and ML libraries
openai>=1.12.0
pinecone>=5.0.0

# PDF processing
PyPDF2>=3.0.1

# HTTP requests
requests>=2.31.0

# Async support (usually included in Python 3.7+)
# asyncio is built-in, but including aiohttp for advanced async HTTP if needed
aiohttp>=3.9.0

# Logging and utilities (built-in, but good to specify versions)
typing-extensions>=4.8.0

# Environment variable loading
python-dotenv>=1.0.0

# Numerical computing for conflict detection
numpy>=1.24.0

# Production API dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# PostgreSQL async database
asyncpg>=0.29.0

# Optional: For better PDF processing if PyPDF2 has issues
# pypdf>=3.0.0
# pdfplumber>=0.10.0

# Optional: For advanced text processing
# nltk>=3.8
# spacy>=3.7.0

# Optional: For better async performance
# uvloop>=0.19.0  # Unix only
